// client/src/app/stores/jobPosts/jobPosts.types.ts
export enum JOB_POSTS_ACTION_TYPES {
    FETCH_ALL_JOB_POSTS_LOADING = "jobPosts/FETCH_ALL_JOB_POSTS_LOADING",
    FETCH_ALL_JOB_POSTS_SUCCESS = "jobPosts/FETCH_ALL_JOB_POSTS_SUCCESS",
    FETCH_ALL_JOB_POSTS_FAILED = "jobPosts/FETCH_ALL_JOB_POSTS_FAILED",
    FETCH_JOB_POST_BY_ID_LOADING = "jobPosts/FETCH_JOB_POST_BY_ID_LOADING",
    FETCH_JOB_POST_BY_ID_SUCCESS = "jobPosts/FETCH_JOB_POST_BY_ID_SUCCESS",
    FETCH_JOB_POST_BY_ID_FAILED = "jobPosts/FETCH_JOB_POST_BY_ID_FAILED",
    ADD_JOB_POST_LOADING = "jobPosts/ADD_JOB_POST_LOADING",
    ADD_JOB_POST_SUCCESS = "jobPosts/ADD_JOB_POST_SUCCESS",
    ADD_JOB_POST_FAILED = "jobPosts/ADD_JOB_POST_FAILED",
    UPDATE_JOB_POST_LOADING = "jobPosts/UPDATE_JOB_POST_LOADING",
    UPDATE_JOB_POST_SUCCESS = "jobPosts/UPDATE_JOB_POST_SUCCESS",
    UPDATE_JOB_POST_FAILED = "jobPosts/UPDATE_JOB_POST_FAILED",
    DELETE_JOB_POST_LOADING = "jobPosts/DELETE_JOB_POST_LOADING",
    DELETE_JOB_POST_SUCCESS = "jobPosts/DELETE_JOB_POST_SUCCESS",
    DELETE_JOB_POST_FAILED = "jobPosts/DELETE_JOB_POST_FAILED",
    APPLY_TO_JOB_POST_LOADING = "jobPosts/APPLY_TO_JOB_POST_LOADING",
    APPLY_TO_JOB_POST_SUCCESS = "jobPosts/APPLY_TO_JOB_POST_SUCCESS",
    APPLY_TO_JOB_POST_FAILED = "jobPosts/APPLY_TO_JOB_POST_FAILED",
  };